# Add rand sampler for Landau (adapted from ROOT's TRandom)