--- a/src/LandauDistribution.jl
+++ b/src/LandauDistribution.jl
@@ -1,6 +1,6 @@
 module LandauDistribution
 
-import Distributions: location, scale, params, partype, pdf, cdf, cf
+import Distributions: location, scale, params, partype, pdf, cdf, cf, rand, AbstractRNG
 using Distributions: @check_args, @distr_support, ContinuousUnivariateDistribution
 export Landau