Hi,

I just very quickly added the sampler for the Landau distribution as is implemented in: https://root.cern/root/html524/src/TRandom.cxx.html#332 

It can probably be optimized a bit, but seems to work for now.

Here is a little script for comparing the resulting random samples with the pdf:
```
using LandauDistribution
using Distributions
using Plots
using FHist

μ = 1.5
θ = 0.2
x = 0:0.1:10
ld = Landau(μ, θ)

r = rand(ld, Int(1e6))
h1 = Hist1D(r, x)

plot(normalize(h1))
plot!(x, pdf.(ld, x), lw=3, label="pdf(Landau($(μ), $(θ))")

```
[plot.pdf](https://github.com/JuliaHEP/LandauDistribution.jl/files/11556062/plot.pdf)