@Moelf 2023-05-24
```julia
### AFTER
julia> @benchmark rand(L)
BenchmarkTools.Trial: 10000 samples with 595 evaluations.
 Range (min … max):  199.321 ns … 270.363 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     203.345 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   204.037 ns ±   2.631 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

            ▂▁▂▅▇█▇▅▃▄▅▃▁       ▁ ▁                             ▂
  ▅▄▅▂▄▅▄▄▆██████████████▆▅▅▃▅▆▇██████▇██▇▇▇▇▇▇▆▇▆▇▇▇▆▆▆▄▅▄▅▅▂▄ █
  199 ns        Histogram: log(frequency) by time        215 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

### BEFORE
julia> @benchmark rand(L)
BenchmarkTools.Trial: 10000 samples with 171 evaluations.
 Range (min … max):  661.433 ns …   6.760 μs  ┊ GC (min … max):  0.00% … 84.08%
 Time  (median):     703.386 ns               ┊ GC (median):     0.00%
 Time  (mean ± σ):   873.037 ns ± 927.875 ns  ┊ GC (mean ± σ):  17.84% ± 14.42%

  █▄                                                          ▂ ▁
  ██▇▆▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁██ █
  661 ns        Histogram: log(frequency) by time       6.17 μs <

 Memory estimate: 7.81 KiB, allocs estimate: 1.
```