@Cornelius-G 2023-05-24
Ah, right. Thanks a lot!👍 Had to run for a meeting before doing some benchmarks. But of course I should have thought of using SVector directly when seeing this large constant array 🙈